requirements.txt
absl-py==1.4.0
cachetools==5.3.1
click==8.0.4
Flask==2.2.2
future==0.18.3
gast==0.4.0
grpcio==1.58.0
gunicorn==20.0.4
h5py==3.7.0
idna==3.4
itsdangerous==2.0.1
Jinja2==3.1.2
Markdown==3.4.1
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.8.1.78
Pillow==9.4.0
protobuf==4.24.3
PyYAML==6.0
requests==2.28.1
scipy==1.10.0
six==1.16.0
#tensorboard==2.13.0
tensorboard==2.8.0
termcolor==2.3.0
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.207
urllib3==1.26.14
Werkzeug==2.2.2
tensorflow-cpu==2.8.0